+++ core/modules/modLIMS.class.php
@@ -186,7 +186,27 @@ class modLIMS extends DolibarrModules
 		// 'user'             to add a tab in user view
 
 		// Dictionaries
-		$this->dictionaries = array();
+		$this->dictionaries = array(
+			'langs'=>'lims@lims',
+			// List of tables we want to see into dictionary editor
+			'tabname'=>array(MAIN_DB_PREFIX."lims_units"),
+			// Label of tables
+			'tablib'=>array("LIMS Units"),
+			// Request to select fields
+			'tabsql'=>array('SELECT f.rowid as rowid, f.code, f.label, f.short_label, f.active FROM '.MAIN_DB_PREFIX.'lims_units as f'),
+			// Sort order
+			'tabsqlsort'=>array("label ASC"),
+			// List of fields (result of select to show dictionary)
+			'tabfield'=>array("code,label,short_label"),
+			// List of fields (list of fields to edit a record)
+			'tabfieldvalue'=>array("code,label,short_label"),
+			// List of fields (list of fields for insert)
+			'tabfieldinsert'=>array("code,label,short_label"),
+			// Name of columns with primary key (try to always name it 'rowid')
+			'tabrowid'=>array("rowid"),
+			// Condition to show each dictionary
+			'tabcond'=>array($conf->lims->enabled)
+		);
 		/* Example:
 		$this->dictionaries=array(
 			'langs'=>'lims@lims',